docker/app/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
prometheus-flask-exporter==0.23.1
